Move implementation into extensions
