+++ jvm/sargon-android/src/main/java/com/radixdlt/sargon/extensions/Accounts.kt
@@ -2,8 +2,6 @@ package com.radixdlt.sargon.extensions
 
 import com.radixdlt.sargon.Account
 import com.radixdlt.sargon.AccountAddress
-import com.radixdlt.sargon.IdentifiedArray
-import com.radixdlt.sargon.IdentifiedArrayImpl
 
 class Accounts private constructor(
     array: IdentifiedArray<AccountAddress, Account>
