+++ jvm/sargon-android/src/main/java/com/radixdlt/sargon/extensions/FactorSources.kt
@@ -2,8 +2,6 @@ package com.radixdlt.sargon.extensions
 
 import com.radixdlt.sargon.FactorSource
 import com.radixdlt.sargon.FactorSourceId
-import com.radixdlt.sargon.IdentifiedArray
-import com.radixdlt.sargon.IdentifiedArrayImpl
 
 class FactorSources private constructor(
     array: IdentifiedArray<FactorSourceId, FactorSource>
