+++ jvm/sargon-android/src/main/java/com/radixdlt/sargon/extensions/ProfileNetworks.kt
@@ -1,7 +1,5 @@
 package com.radixdlt.sargon.extensions
 
-import com.radixdlt.sargon.IdentifiedArray
-import com.radixdlt.sargon.IdentifiedArrayImpl
 import com.radixdlt.sargon.NetworkId
 import com.radixdlt.sargon.ProfileNetwork
